third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-basics.html

<!DOCTYPE html>
<title>Credential Manager: CredentialsContainer basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function () {
    assert_equals(typeof CredentialsContainer, "function");
    verifyInterface('CredentialsContainer', navigator.credentials, {
        'request': 'function'
    });
}, 'Interfaces and attributes of CredentialsContainer');
</script>
